backend/requirements.txt
fastapi==0.116.1
python-multipart==0.0.20
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
python-dotenv==1.1.1
openpyxl==3.1.5
pytest==8.4.2
httpx==0.28.1
